♻️ refactor: YYYY.MM.DD 형식으로 포메팅하는 함수 분리, 재사용, 마우스 위치 함수 주석 처리 #123 날짜 형식을 포메팅하는 함수를 PriceTableDayCard.tsx에서 분리 후 재사용. 켄버스 내에서 마우스 위치를 찾는 함수를 아직 사용하지 않기 때문에 주석 처리.
