+++ FE/src/components/StocksDetail/PriceTableDayCard.tsx
@@ -1,4 +1,5 @@
 import { DailyPriceDataType } from './PriceDataType.ts';
+import { formatTime } from '../../utils/formatTime.ts';
 
 type PriceTableDayCardProps = {
   data: DailyPriceDataType;
@@ -15,13 +16,6 @@ export default function PriceTableDayCard({ data }: PriceTableDayCardProps) {
 
   const plusOrMinus =
     data.prdy_vrss_sign === '3' ? '' : data.prdy_vrss_sign < '3' ? '+' : '-';
-  function formatTime(time: string) {
-    if (!time.length) return '----.--.--';
-    const year = time.slice(0, 4);
-    const mon = time.slice(4, 6);
-    const day = time.slice(6, 8);
-    return `${year}.${mon}.${day}`;
-  }
   return (
     <tr className={'h-[30px] hover:bg-juga-grayscale-50'}>
       <td className={'px-4 py-1 text-start'}>
