+++ FE/src/utils/formatTime.ts
@@ -1,0 +1,7 @@
+export function formatTime(time: string) {
+  if (!time.length) return '----.--.--';
+  const year = time.slice(0, 4);
+  const mon = time.slice(4, 6);
+  const day = time.slice(6, 8);
+  return `${year}.${mon}.${day}`;
+}
